classes were missing public declaration
