+++ java/libthither/src/main/java/direct/thither/lib/api/fms/FmsDefinitionType.java
@@ -1,0 +1,8 @@
+package direct.thither.lib.api.fms;
+/*
+ * Author Kashirin Alex (kashirin.alex@gmail.com)
+ * THITHER.DIRECT
+ * */
+
+
+public enum FmsDefinitionType{ALL, UNITS, SECTIONS, METRICS}
